DAOS-17914 dlck: move daos_srv/dlck.h -> daos/dlck.h
Skip-func-hw-test: true
